pyproject.toml
[tool.poetry]
name = "it-park-backend"
version = "0.1.0"
description = ""
authors = ["Askhat <askhat.ss23@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.11"
uvicorn = "^0.34.0"
pymongo = "^4.11.3"
email-validator = "^2.2.0"
pydantic = "^2.10.6"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
